+++ protos/embed.go
@@ -1,0 +1,9 @@
+package protos
+
+import (
+	"embed"
+	_ "embed"
+)
+
+//go:embed generated
+var Generated embed.FS